Only flush USB printer commands when a whole INonFiscalDevice method is executed
Due to the early closing of the USB Device, some errors may have ocurred. On
a given situation, garbage may have been read by the printer on the middle
of a series of commands.

```
self.write(...)
self.write(...)
self.write(...)

self.write(...)
...
```

As we do not have the possession of the USB Resource between the `self.write`s
(and we do not know if the printer still accepts commands after the
interface has been closed), some commands can be corrupted.

Change-Id: I1ab0295bdb0686af38b7df3ce1648b69726c18e5