Send params sorted to FiscNetECF
Python doesn't guarantee the order of dicts, it is just that in
python2, because of the way they are implemented, they usually will
keep their order when iterating over them.

In python3 the dict implementation changed in a way that the order is
really unpredictable.

This change will sort the params so their order will be deterministic and
the tests will be consistent.

Change-Id: Ic7af3c755504fe95191b78d28b9cf5631cab1aef