Fix issue introduced by my last refactor.
We can't keep opening/closing the device, otherwise it will get really slot
to print (as reported by a user).

This is moving the printer setup to each printer constructor, leaving it out
for the usb ones (they didn't have those before the refactor anyway). And
leaving the device open, as it was before.

Changing this will need a bigger refactor of the driver and callsites to
work properly.

Change-Id: I27778e4bcc0d4cc334bec80ea678b0879bbdd30b