Implement config for cash drawers with inverted logic
There are specific models of cash drawers (Radiant, for instance) where the
logic that determines whether it is open is inverted; in other words, if for
standard drawers the return value that indicates "drawer open" is X, for
inverted drawers, it's !X.

This change allows specifying inverted_drawer=True through config and the
BaseDevice constructor. The default setting is False.

The configuration code is a bit overcomplicated but I've tried to preserve the
original semantics as much as possible; the configfile can however now be used
to override configuration specified in the constructor. I felt that was the
most useful option.

I've changed the configuration file path to be use "stoq" instead of
"stoqdrivers"; this is incompatible and may break existing configurations
but I think the risk is low and I want to avoid one more directory. The
filename, stoqdrivers.conf, should be unique enough.

This change adds testing for drawer open and drawer check calls.

I've also bumped the version in anticipation of a new release.

Fixes: https://gitlab.com/stoqtech/private/bdil/issues/86
Change-Id: I16e68bd412ce1bbd3a7b9f5aaef8738cc7d27e1c