Adjust text encoding handling for non fiscal printers
NonFiscalPrinter wrapper might receive strings or bytes. In case it
receives strings, it will be encoded into bytes using the encoding
defined in the driver implementation (or unicode if not defined by the
driver).

So, NonFiscalPrinter will always pass bytes to the driver implementation
(which is expected by pyserial).

This fix adjusts the text encoding of the Bematech MP2100TH and
MP4200TH printers so that non-ascii characters can be printed whenever
needed.

Fixes: https://gitlab.com/stoqtech/private/bdil/issues/260
Change-Id: I316200c5adec9bbcc68bb5f3dc722158f4fa98ae